crm-platform/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
streamlit==1.37.1
sqlalchemy==2.0.23
requests==2.31.0
plotly==5.0.0
python-multipart==0.0.6
pydantic==1.10.12
python-dotenv==0.21.0
aiofiles==23.2.1
httpx==0.23.3
